refactor(action-log): extract db persistence into Action_Log_Store
